+++ src/index.js
@@ -7,7 +7,6 @@
 
 const pkg = require('../package.json');
 const Wurst = require('./wurst');
-const factory = require('./factory');
 require('babel-polyfill');
 
 /**
@@ -20,10 +19,7 @@ require('babel-polyfill');
  * @returns {*}
  */
 function pluginLoader(server, options, next) {
-  const test = Wurst.init(server, options)
-  test.load();
-
-  console.log(factory.isPrototypeOf(test));
+  Wurst.init(server, options).load();
 
   return next();
 }